[mutable] Implement memory budget control for image operations. Cleanup memory budget interface. Remove unused image operations. #prefight 646c958c4422ba05f47ea1e6 #rnx [REVIEW] [at]genis.sole [REVIEW] [at]gerard.martin [REVIEW] [at]pere.rifa [REVIEW] [at]alexei.lebedev
[CL 25598403 by jordi rovira in ue5-main branch]

+++ Engine/Plugins/Experimental/Mutable/Source/MutableRuntime/Private/MuR/MeshBufferSet.cpp
@@ -380,9 +380,9 @@ namespace mu
 
 
 	//---------------------------------------------------------------------------------------------
-	size_t FMeshBufferSet::GetDataSize() const
+	int32 FMeshBufferSet::GetDataSize() const
 	{
-		size_t res = 0;
+		int32 res = 0;
 
 		for (int32 b = 0; b < m_buffers.Num(); ++b)
 		{
